STAFF-46 Added permissons on endpoints

+++ Modules/Tips/Http/Controllers/api/CategoryController.php
@@ -2,9 +2,8 @@
 
 namespace Modules\Tips\Http\Controllers\api;
 
-use Illuminate\Contracts\Support\Renderable;
-use Illuminate\Http\Request;
-use Illuminate\Routing\Controller;
+use App\Http\Controllers\Controller;
+use Illuminate\Auth\Access\AuthorizationException;
 use Modules\Tips\Entities\Category;
 use Modules\Tips\Http\Requests\CategoryRequest;
 use Modules\Tips\Transformers\CategoryResource;
@@ -16,24 +15,39 @@ class CategoryController extends Controller
         return CategoryResource::collection(Category::all());
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(CategoryRequest $request): CategoryResource
     {
+        $this->authorize('administration', Auth()->user());
+
         return CategoryResource::make(Category::create([
             'name' => $request->name,
             'slug' => str_slug($request->name, '_'),
         ]));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(Category $category, CategoryRequest $request): CategoryResource
     {
+        $this->authorize('administration', Auth()->user());
+
         return CategoryResource::make($category->update([
             'name' => $request->name,
             'slug' => str_slug($request->name, '_'),
         ]));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Category $category): \Illuminate\Http\Resources\Json\AnonymousResourceCollection
     {
+        $this->authorize('administration', Auth()->user());
+
         $category->delete();
         return CategoryResource::collection(Category::all());
     }
